+++ include/Sprite.h
@@ -9,10 +9,12 @@
 #define _SPRITE_H_
 
 #include "math/math.h"
+#include "TombRaider.h"
 
 class Sprite {
 public:
-    Sprite(vec3_t _vertex[4], vec2_t _texel[4], vec3_t _pos, vec_t _radius, int _texture);
+    Sprite(TombRaider &tr, unsigned int room, unsigned int index);
+    Sprite(TombRaider &tr, unsigned int item, unsigned int sequence, unsigned int index);
     void display();
 
 private:
@@ -25,6 +27,7 @@ private:
 
 class SpriteSequence {
 public:
+    SpriteSequence(TombRaider &tr, unsigned int item, unsigned int sequence);
     ~SpriteSequence();
 
     void add(Sprite &s);