+++ ChangeLog.md
@@ -2,6 +2,10 @@
 
 ## OpenRaider (0.1.3) xythobuz <xythobuz@xythobuz.de>
 
+    [ 20140505 ]
+    * Moved setup of Room, Portal, Box, Sector, ... into their
+      Constructors.
+
     [ 20140504 ]
     * Forcing use of new Room and Sprite classes in World.
     * Rewrote Sprite loading code in Game